% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmd_skeleton.R
\name{.jmv_asstring_to_html}
\alias{.jmv_asstring_to_html}
\title{Parse a jmv Table$asString() ASCII layout into a clean HTML table string.}
\usage{
.jmv_asstring_to_html(s)
}
\description{
Parse a jmv Table$asString() ASCII layout into a clean HTML table string.
}
\keyword{internal}
