% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Value to check.}

\item{y}{Default value to return if \code{x} is \code{NULL}.}
}
\value{
\code{x} if not \code{NULL}, otherwise \code{y}.
}
\description{
Returns the left-hand side if it is not \code{NULL}, otherwise
returns the right-hand side. This is a convenience operator used throughout
the converter packages for default value handling.
}
\details{
Null-coalescing operator
}
\examples{
# Returns the non-NULL value
"hello" \%||\% "default"

# Returns the default when NULL
NULL \%||\% "default"
}
