% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat2rmarkdown-package.R
\docType{package}
\name{stat2rmarkdown-package}
\alias{stat2rmarkdown}
\alias{stat2rmarkdown-package}
\title{stat2rmarkdown: Shared Core Utilities for Statistical Software to R Markdown Converters}
\description{
Common utilities for converting statistical software output files to reproducible R Markdown reports. Provides R Markdown document assembly, variable quoting, output directory management, and logging utilities shared across the stata2rmarkdown, spss2rmarkdown, jasp2rmarkdown, and jamovi2rmarkdown packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giladfeldman/stat2rmarkdown}
  \item Report bugs at \url{https://github.com/giladfeldman/stat2rmarkdown/issues}
}

}
\author{
\strong{Maintainer}: Gilad Feldman \email{giladfel@gmail.com} (\href{https://orcid.org/0000-0003-2990-8997}{ORCID})

Authors:
\itemize{
  \item Gilad Feldman \email{giladfel@gmail.com} (\href{https://orcid.org/0000-0003-2990-8997}{ORCID})
}

}
\keyword{internal}
